bot/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
py-cord>=2.4.1
colorama>=0.4.6
pytz>=2023.3
humanize>=4.8.0
Pillow>=10.0.0
requests>=2.31.0
